+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Next
+
+- fix: display proper error message when Playwright is not installed (#10)
+
 ## v0.3.0 (2025-06-04)
 
 - feat: added SSE transport; you can now use this with Claude Code 🤖